refactor: split track and page tracker
